extract tortoise
